backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.0.1
pillow==10.1.0
aiofiles==23.2.1
websockets==12.0
httpx==0.25.2
reportlab==4.0.8
